requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.36.0
requests==2.32.3
pymupdf
python-docx
pandas
chromadb
python-multipart
